requirements.txt
gradio==3.50.2
numpy
Pillow
tqdm
sqlalchemy
matplotlib
scikit-learn
ipykernel
